[admin][sr] refacto members section into 2 components
